+++ app/lib/csvMappers/sparkasse.ts
@@ -1,0 +1,5 @@
+import { mapGermanBank } from "./germanBank";
+
+export function mapSparkasse(row: Record<string, string>) {
+  return mapGermanBank(row, "sparkasse");
+}
