+++ app/lib/csvMappers/types.ts
@@ -1,0 +1,11 @@
+export interface TransactionData {
+  date: number;
+  amount: number;
+  description: string;
+  counterparty: string;
+  bankReferenceId: string;
+  accountName?: string;
+  currency?: string;
+}
+
+export type ImportSource = "moss" | "sparkasse" | "volksbank" | "finom";
