+++ app/lib/csvMappers/volksbank.ts
@@ -1,0 +1,5 @@
+import { mapGermanBank } from "./germanBank";
+
+export function mapVolksbank(row: Record<string, string>) {
+  return mapGermanBank(row, "volksbank");
+}
